refactor: ManagerJoin 컴포넌트와 JoinForm 컴포넌트의 관심사 분리
