+++ frontend/src/pages/ManagerJoin/ManagerJoin.tsx
@@ -1,18 +1,50 @@
+import { AxiosError } from 'axios';
+import React from 'react';
+import { useMutation } from 'react-query';
+import { useHistory } from 'react-router';
 import { Link } from 'react-router-dom';
+import { postJoin } from 'api/join';
 import Header from 'components/Header/Header';
 import Layout from 'components/Layout/Layout';
+import MESSAGE from 'constants/message';
 import PATH from 'constants/path';
+import { ErrorResponse } from 'types/response';
 import * as Styled from './ManagerJoin.styles';
 import JoinForm from './units/JoinForm';
 
+export interface JoinParams {
+  email: string;
+  password: string;
+  organization: string;
+}
+
 const ManagerJoin = (): JSX.Element => {
+  const history = useHistory();
+
+  const join = useMutation(postJoin, {
+    onSuccess: () => {
+      alert(MESSAGE.JOIN.SUCCESS);
+      history.push(PATH.MANAGER_LOGIN);
+    },
+
+    onError: (error: AxiosError<ErrorResponse>) => {
+      alert(error?.response?.data.message ?? MESSAGE.JOIN.FAILURE);
+    },
+  });
+
+  const handleSubmit = ({ email, password, organization }: JoinParams) => {
+    if (!email || !password || !organization) return;
+
+    join.mutate({ email, password, organization });
+  };
+
   return (
     <>
       <Header />
       <Layout>
         <Styled.Container>
           <Styled.PageTitle>회원가입</Styled.PageTitle>
-          <JoinForm />
+          <JoinForm onSubmit={handleSubmit} />
           <Styled.JoinLinkMessage>
             이미 회원이신가요?
             <Link to={PATH.MANAGER_LOGIN}>로그인하기</Link>
